[myagv][code] Add a moving condition
In the previous codes, the position information was gain from the tf.transform. When the /tag_detections information is none, the tf will publish the last one forever. So the agv will move forward with the same velocity. Now I add a moving condition. Only the /tag_detections is not none, the agv can move.