[Qilin][script] cannot exchange the /quadrotor/uav/nav topic
It is wired yesterday and today the /quadrotor/uav/nav cannot publish in the drone. The dog published but drone cannot get that. What's more, the topic with msg like "Pose"/"FlightNav". "Empty"or"Unit8" is ok.

So I use the way which is not smart. Use /quadrotor/uav/nav/trigger in low speed way as trigger and use /quadrotor/uav/nav/info in highway to give the pose.